test(msteams): split normalization authz coverage
